asm decompiler: move resolveType to AsmReferenceResolver
